Offloaded some of the Edge fields to newly added ActiveEdge.
This was a major rework which involved migration of Edge.(kind/side/contributing/poly) fields to new class ActiveEdge. ActiveEdge is a replacement of DoublyList <Edge> which aims to hold some of the active edge data which is not necessary to be in Edge. This saves a lot of memory. Previously instances of the Edge class needed 48 bytes of memory, now they need 32 bytes which is 1.5 times less. It's notable that aj clipper requires 108 (!) bytes per edge which is 3.375 times larger than current size of my edges!
Of course fields that were offloaded are not just thrown away. They are now simply in the "right" place, in active edge list. In fact no other code really needed them.
Beside migrated fields ActiveEdge class is very similiar to DoublyList <Edge>. The only notable difference is in insertNext/insertPrev: insertion creates new instance of ActiveEdge instead of DoublyList <Edge>.

Some other classes were affected too. These are:
1. Intersection.hx - now it contain pointers to ActiveEdge instead of DoublyList <Edge>.
2. LocalMaxima.hx - it has new field kind:PolyKind which is used upon ActiveEdge creation.