X-intercepts with the top of the scanbeam are now cached; TopX is now calculated avoiding fp-accumulation errors.
Edge.bottomX migrated to ActiveEdge.bottomXIntercept. Now Edge.bottomX have slightly different meaning: it's just an X-coordinate of bottom vertex even when edge was added to AEL. In addition to bottomXIntercept ActiveEdge now have two other fields: bottomY and topXIntercept. First is like Edge.bottomX except that it doesn't need to be in Edge because we have y's in local minima list and therefore we can save memory by storing this value only in active edges. Second is above mentioned cache for the x-intercept with the top of the scanbeam.