Added processing of intersections lying exactly in local minimas which previously were unhandled.
Modified VattiClipper.processIntersections () so that now it finds intersections lying in local minimas after finding "ordinary" intersections. Algorithm adds these intersections with k set to positive infinity because they're last intersections in the scanbeam. Right bound choosen to intersect all edges passing through local minima so it is possible that right bound will be intersected twice by some edge(s) when rightBoundEdge.topXIntercept calculated slightly righter than real x-value of the local minima. But this (very rare) inaccurate behavior doesn't breaks anything.
Also ilLast was added: this variable tracks last intersection in intersection list which allows fast insertion of new elements to the end of the list. By the way, there is a suspicion that tracking of ilLast will slowdown noticeably process of filling intersection list and that it is better to add local-minima-intersections by ordinary means. Profile it in the future.