I think I solved problem with too much close Intersection.k's in intersection list which caused wrong sorting order.
Modified VattiClipper.hx processIntersectionList (): now it checks whether intersection e1Node and e2Node currently adjacent in Active Edge List. If not, then it finds isec with that property evaluated to true. After all it removes isec from Intersection List and takes first intersection as next in processing. This guarantees (*) that in the end of processIntersectionList (), ordering of Active Edge List will be same as when Intersection.k's were calculated precisely.

(*) it requires rigid proof but I have serious reason to believe that it's true.