VattiClipper.hx initLmlAndSbl (): determination of horizontal edges now more rigid.
It was possible that k = dx / dy give infinite k which means that edge is very close to be horizontal. For example valid dx = 2^1023 divided by valid dy = 0.5 will give us positive infinity and we can't use it in future calculations. However it is clear that edge with those dx and dy is indistinguishable from horizontal so we just can set dy to zero and proceed as if it was horizontal initially.