Improved accuracy of VattiClipper.intersectionOf ().
As previously it finds Y-coordinate of intersection first and then calculates X-coordinate based on Y. But there is one change: before calculating X it looks which edge has lesser abs ( dx ) and chooses that edge to participate in calculations. Rationale for this: VattiClipper.topX () gives more accurate result for less-horizontal edge.