Bugfix for BottomUpEdge.topX () called on horizontal edge.
Added OutputBound.topX () which is a proxy between caller and BottomUpEdge.topX (). It checks whether edge is horizontal and if so, returns X-coordinate of the last point added to the edge, otherwise it calls BottomUpEdge.topX (). This approach prevents us from returning unwanted NaN from topX () called on horizontal edge.