Bugfix in PolyBounds.initLml ().
Bug occured when all points of an input poly were on the same horizontal line. All of them considered decreasing except of the last edge which is considered increasing. So, at the point in code where last edge was constructed by connecting first and last points, i forgot to set its isHorizontal field to true.