Added PolyBounds.swapKinds () method.
This method loops through local minimas stored in given instance of PolyBounds and changes their kind field to opposite value. It is helpful when we want to reuse same PolyBounds instance for different clipping operations. For example we subtract some geometry from PolyBounds object A first. In order to do this, we must construct A with its kind set to PolyKind.Subject. And then, if we want to subtract A from some other geometry, we had to construct A with its kind set to PolyKind.Clip. This is where PolyBounds.swapKinds () comes into play: we can just swap kinds and reuse this PolyBounds object in second subtraction.